improve server API for later publishing
