+++ discojs/discojs-core/src/dataset/data/tabular_data.spec.ts
@@ -1,18 +1,49 @@
 import { assert, expect } from 'chai'
+import { Map, Set } from 'immutable'
 
-import { tasks, node } from '@epfml/discojs-node'
+import { TabularData } from './tabular_data'
+import { tf, Task } from '../..'
 
 describe('tabular data checks', () => {
-  const task = tasks.titanic.task
+  const titanicMock: Task = {
+    taskID: 'titanic',
+    displayInformation: {},
+    trainingInformation: {
+      inputColumns: [
+        'PassengerId',
+        'Age',
+        'SibSp',
+        'Parch',
+        'Fare',
+        'Pclass'
+      ],
+      outputColumns: [
+        'Survived'
+      ]
+    }
+  } as unknown as Task
+
   const dataConfig = {
-    features: task.trainingInformation.inputColumns,
-    labels: task.trainingInformation.outputColumns
+    features: titanicMock.trainingInformation.inputColumns,
+    labels: titanicMock.trainingInformation.outputColumns
   }
-  const loader = new node.data.NodeTabularLoader(task, ',')
+
+  const columnConfigs = Map(
+    Set(dataConfig.features).map((feature) => [feature, { required: false, isLabel: false }])
+  ).merge(
+    Set(dataConfig.labels).map((label) => [label, { required: true, isLabel: true }])
+  )
+
+  const csvConfig = {
+    hasHeader: true,
+    columnConfigs: columnConfigs.toObject(),
+    configuredColumnsOnly: true,
+    delimiter: ','
+  }
 
   it('throw an error on incorrectly formatted data', async () => {
     try {
-      await loader.loadAll(['../../example_training_data/cifar10-labels.csv'], dataConfig)
+      await TabularData.init(tf.data.csv('file://../../example_training_data/cifar10-labels.csv', csvConfig), titanicMock, 3)
     } catch (e) {
       expect(e).to.be.an.instanceOf(Error)
       return
@@ -22,6 +53,6 @@ describe('tabular data checks', () => {
   })
 
   it('do nothing on correctly formatted data', async () => {
-    await loader.loadAll(['../../example_training_data/titanic_train.csv'], dataConfig)
+    await TabularData.init(tf.data.csv('file://../../example_training_data/titanic_train.csv', csvConfig), titanicMock, 3)
   })
 })
