+++ discojs/discojs-core/src/default-tasks/index.ts
@@ -1,0 +1,6 @@
+export { cifar10 } from './cifar10'
+export { lusCovid } from './lus_covid'
+export { mnist } from './mnist'
+export { titanic } from './titanic'
+export { simpleFace } from './simple_face'
+export { geotags } from './geotags'
