+++ discojs/discojs-core/src/serialization/model.spec.ts
@@ -1,6 +1,6 @@
 import { assert } from 'chai'
 
-import { tf, tasks, serialization } from '@epfml/discojs-node'
+import { tf, serialization } from '@epfml/discojs-node'
 
 async function getRawWeights (model: tf.LayersModel): Promise<Array<[number, Float32Array]>> {
   return Array.from(
@@ -12,7 +12,16 @@ async function getRawWeights (model: tf.LayersModel): Promise<Array<[number, Flo
 
 describe('model', () => {
   it('can encode what it decodes', async () => {
-    const model = tasks.mnist.model()
+    const model = tf.sequential()
+
+    model.add(
+      tf.layers.conv2d({
+        inputShape: [32, 32, 3],
+        kernelSize: 3,
+        filters: 16,
+        activation: 'relu'
+      })
+    )
 
     const encoded = await serialization.model.encode(model)
     assert.isTrue(serialization.model.isEncoded(encoded))
