+++ discojs/discojs-core/src/task/index.ts
@@ -1,3 +1,5 @@
 export { isTask, Task, isTaskID, TaskID } from './task'
+export { TaskProvider, isTaskProvider } from './task_provider'
+export { isDigest, Digest } from './digest'
 export { isDisplayInformation, DisplayInformation } from './display_information'
 export { TrainingInformation } from './training_information'
