--- discojs/discojs-core/src/tasks/index.ts
@@ -1,6 +1,0 @@
-export * as cifar10 from './cifar10'
-export * as lus_covid from './lus_covid'
-export * as mnist from './mnist'
-export * as titanic from './titanic'
-export * as simple_face from './simple_face'
-export * as geotags from './geotags'
