+++ discojs/discojs-node/src/dataset/data_loader/image_loader.spec.ts
@@ -2,7 +2,7 @@ import { assert, expect } from 'chai'
 import { List, Map, Range } from 'immutable'
 import fs from 'fs'
 
-import { tf, tasks, node } from '@epfml/discojs-node'
+import { tf, node, Task } from '../../..'
 
 const readFilesFromDir = (dir: string): string[] =>
   fs.readdirSync(dir).map((file: string) => dir + file)
@@ -10,9 +10,30 @@ const readFilesFromDir = (dir: string): string[] =>
 const DIRS = {
   CIFAR10: '../../example_training_data/CIFAR10/'
 }
+
+const cifar10Mock: Task = {
+  taskID: 'cifar10',
+  displayInformation: {},
+  trainingInformation: {
+    IMAGE_H: 32,
+    IMAGE_W: 32,
+    LABEL_LIST: ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9']
+  }
+} as unknown as Task
+
+const mnistMock: Task = {
+  taskID: 'mnist',
+  displayInformation: {},
+  trainingInformation: {
+    IMAGE_H: 28,
+    IMAGE_W: 28,
+    LABEL_LIST: ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9']
+  }
+} as unknown as Task
+
 const LOADERS = {
-  CIFAR10: new node.data.NodeImageLoader(tasks.cifar10.task),
-  MNIST: new node.data.NodeImageLoader(tasks.mnist.task)
+  CIFAR10: new node.data.NodeImageLoader(cifar10Mock),
+  MNIST: new node.data.NodeImageLoader(mnistMock)
 }
 const FILES = Map(DIRS).map((readFilesFromDir)).toObject()
 
@@ -69,7 +90,7 @@ describe('image loader', () => {
   })
 
   it('loads samples in order', async () => {
-    const loader = new node.data.NodeImageLoader(tasks.cifar10.task)
+    const loader = new node.data.NodeImageLoader(cifar10Mock)
     const dataset = await ((await loader.loadAll(FILES.CIFAR10, { shuffle: false })).train.dataset).toArray()
 
     List(dataset).zip(List(FILES.CIFAR10))
@@ -83,7 +104,7 @@ describe('image loader', () => {
   })
 
   it('shuffles list', async () => {
-    const loader = new node.data.NodeImageLoader(tasks.cifar10.task)
+    const loader = new node.data.NodeImageLoader(cifar10Mock)
     const list = Range(0, 100_000).toArray()
     const shuffled = [...list]
 
@@ -95,7 +116,7 @@ describe('image loader', () => {
   })
 
   it('shuffles samples', async () => {
-    const loader = new node.data.NodeImageLoader(tasks.cifar10.task)
+    const loader = new node.data.NodeImageLoader(cifar10Mock)
     const dataset = await (await loader.loadAll(FILES.CIFAR10, { shuffle: false })).train.dataset.toArray()
     const shuffled = await (await loader.loadAll(FILES.CIFAR10, { shuffle: true })).train.dataset.toArray()
 
@@ -107,7 +128,7 @@ describe('image loader', () => {
   it('validation split', async () => {
     const validationSplit = 0.2
     const imagesContent = FILES.CIFAR10.map((file) => tf.node.decodeImage(fs.readFileSync(file)))
-    const datasetContent = await new node.data.NodeImageLoader(tasks.cifar10.task)
+    const datasetContent = await new node.data.NodeImageLoader(cifar10Mock)
       .loadAll(FILES.CIFAR10, { shuffle: false, validationSplit: validationSplit })
 
     const trainSize = Math.floor(imagesContent.length * (1 - validationSplit))
