+++ server/src/index.ts
@@ -1,1 +1,6 @@
-export { getApp } from './get_server'
+import { tf } from '@epfml/discojs-node'
+
+export { Disco } from './get_server'
+
+// TODO, can we let the user of the server retrieve its own tfjs to get the types?
+export { tf }
