+++ server/tests/client/decentralized.spec.ts
@@ -2,11 +2,11 @@ import { expect } from 'chai'
 import { List } from 'immutable'
 import * as http from 'http'
 
-import { tf, client as clients, informant, Task, tasks, WeightsContainer } from '@epfml/discojs-node'
+import { tf, client as clients, informant, Task, WeightsContainer, defaultTasks } from '@epfml/discojs-node'
 
 import { getClient, startServer } from '../utils'
 
-const TASK = tasks.titanic.task
+const TASK = defaultTasks.titanic.getTask()
 
 function test (
   name: string,
